+++ ksa_compliance/ksa_compliance/doctype/zatca_integration_log/zatca_integration_log.py
@@ -6,4 +6,19 @@ from frappe.model.document import Document
 
 
 class ZATCAIntegrationLog(Document):
+	# begin: auto-generated types
+	# This code is auto-generated. Do not modify anything in this block.
+
+	from typing import TYPE_CHECKING
+
+	if TYPE_CHECKING:
+		from frappe.types import DF
+
+		e_invoice_file: DF.Attach | None
+		invoice_additional_fields_reference: DF.Link
+		invoice_reference: DF.Link
+		status: DF.Literal["", "Pending", "Resend", "Accepted with warnings", "Accepted", "Rejected", "Clearance switched off"]
+		zatca_message: DF.LongText | None
+		zatca_status: DF.Data | None
+	# end: auto-generated types
 	pass